Test results, and we are done.

+++ src/main/scala/io/asuna/lucinda/statistics/ResultsGenerator.scala
@@ -99,7 +99,7 @@ object ResultsGenerator {
         rank = index + 1,
         value = value,
         average = average,
-        percentile = 1 - index / statsMap.size
+        percentile = 1 - index.toDouble / statsMap.size
       )
     }
   }
